Improve the SubmitPanel styles
