+++ src/client/components/submit_panel/index.ts
@@ -1,0 +1,1 @@
+export * from './submit_panel';
