+++ src/common/types/events.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent } from "react";
 
 export type InputChangeEvent = ChangeEvent<HTMLInputElement>;
-
+export type SelectChangeEvent = ChangeEvent<HTMLSelectElement>;
 export type TextAreaChangeEvent = ChangeEvent<HTMLTextAreaElement>;
